+++ src/home.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function home(){
+    return(<h1> Ui by Dhruv Gupta</h1>);
+}
+
+export default Home;
